+++ test/libgit2-online.jl
@@ -9,7 +9,8 @@ mktempdir() do dir
     @testset "Cloning repository" begin
         @testset "with 'https' protocol" begin
             repo_path = joinpath(dir, "Example1")
-            repo = LibGit2.clone(repo_url, repo_path)
+            payload = LibGit2.CredentialPayload(allow_prompt=false)
+            repo = LibGit2.clone(repo_url, repo_path, payload=payload)
             try
                 @test isdir(repo_path)
                 @test isdir(joinpath(repo_path, ".git"))
@@ -23,7 +24,7 @@ mktempdir() do dir
                 repo_path = joinpath(dir, "Example2")
                 # credentials are required because github tries to authenticate on unknown repo
                 cred = LibGit2.UserPasswordCredentials("JeffBezanson", "hunter2") # make sure Jeff is using a good password :)
-                payload = LibGit2.CredentialPayload(cred)
+                payload = LibGit2.CredentialPayload(cred, allow_prompt=false)
                 LibGit2.clone(repo_url*randstring(10), repo_path, payload=payload)
                 error("unexpected")
             catch ex
@@ -37,7 +38,7 @@ mktempdir() do dir
                 repo_path = joinpath(dir, "Example3")
                 # credentials are required because github tries to authenticate on unknown repo
                 cred = LibGit2.UserPasswordCredentials("","") # empty credentials cause authentication error
-                payload = LibGit2.CredentialPayload(cred)
+                payload = LibGit2.CredentialPayload(cred, allow_prompt=false)
                 LibGit2.clone(repo_url*randstring(10), repo_path, payload=payload)
                 error("unexpected")
             catch ex