+++ test/libgit2.jl
@@ -1936,26 +1936,23 @@ mktempdir() do dir
             function gen_ex(; username="git")
                 quote
                     include($LIBGIT2_HELPER_PATH)
-                    credential_loop($valid_cred, $url, Nullable{String}($username), use_ssh_agent=true)
+                    payload = CredentialPayload(allow_ssh_agent=true, allow_prompt=false)
+                    credential_loop($valid_cred, $url, Nullable{String}($username), payload)
                 end
             end
 
-            challenges = [
-                "Username for 'github.com':" => "\x04",
-            ]
-
             # An empty string username_ptr
             ex = gen_ex(username="")
-            err, auth_attempts = challenge_prompt(ex, challenges)
-            @test err == abort_prompt  # TODO: `eauth_error` when we can disable prompting
+            err, auth_attempts = challenge_prompt(ex, [])
+            @test err == eauth_error
             @test auth_attempts == 2
 
             # A null username_ptr passed into `git_cred_ssh_key_from_agent` can cause a
             # segfault.
             ex = gen_ex(username=nothing)
-            err, auth_attempts = challenge_prompt(ex, challenges)
-            @test err == abort_prompt  # TODO: `eauth_error` when we can disable prompting
-            @test auth_attempts == 1
+            err, auth_attempts = challenge_prompt(ex, [])
+            @test err == eauth_error
+            @test auth_attempts == 2
         end
 
         @testset "SSH explicit credentials" begin
@@ -1969,22 +1966,22 @@ mktempdir() do dir
             invalid_key = joinpath(KEY_DIR, "invalid")
             invalid_cred = LibGit2.SSHCredentials(username, "", invalid_key, invalid_key * ".pub")
 
-            function gen_ex(cred)
+            function gen_ex(cred; allow_prompt=true)
                 quote
                     include($LIBGIT2_HELPER_PATH)
-                    payload = CredentialPayload($cred)
-                    credential_loop($valid_cred, $url, $username, payload, use_ssh_agent=false)
+                    payload = CredentialPayload($cred, allow_ssh_agent=false, allow_prompt=$allow_prompt)
+                    credential_loop($valid_cred, $url, $username, payload)
                 end
             end
 
             # Explicitly provided credential is correct
-            ex = gen_ex(valid_cred)
+            ex = gen_ex(valid_cred, allow_prompt=true)
             err, auth_attempts = challenge_prompt(ex, [])
             @test err == git_ok
             @test auth_attempts == 1
 
             # Explicitly provided credential is incorrect
-            ex = gen_ex(invalid_cred)
+            ex = gen_ex(invalid_cred, allow_prompt=false)
             err, auth_attempts = challenge_prompt(ex, [])
             @test err == eauth_error
             @test auth_attempts == 2
@@ -1996,22 +1993,22 @@ mktempdir() do dir
             valid_cred = LibGit2.UserPasswordCredentials("julia", randstring(16))
             invalid_cred = LibGit2.UserPasswordCredentials("alice", randstring(15))
 
-            function gen_ex(cred)
+            function gen_ex(cred; allow_prompt=true)
                 quote
                     include($LIBGIT2_HELPER_PATH)
-                    payload = CredentialPayload($cred)
+                    payload = CredentialPayload($cred, allow_prompt=$allow_prompt)
                     credential_loop($valid_cred, $url, "", payload)
                 end
             end
 
             # Explicitly provided credential is correct
-            ex = gen_ex(valid_cred)
+            ex = gen_ex(valid_cred, allow_prompt=true)
             err, auth_attempts = challenge_prompt(ex, [])
             @test err == git_ok
             @test auth_attempts == 1
 
             # Explicitly provided credential is incorrect
-            ex = gen_ex(invalid_cred)
+            ex = gen_ex(invalid_cred, allow_prompt=false)
             err, auth_attempts = challenge_prompt(ex, [])
             @test err == eauth_error
             @test auth_attempts == 2
@@ -2023,11 +2020,11 @@ mktempdir() do dir
 
             valid_username = "julia"
             valid_password = randstring(16)
-            valid_cred = LibGit2.UserPasswordCredentials(valid_username, valid_password, true)
+            valid_cred = LibGit2.UserPasswordCredentials(valid_username, valid_password)
 
             invalid_username = "alice"
             invalid_password = randstring(15)
-            invalid_cred = LibGit2.UserPasswordCredentials(invalid_username, invalid_password, true)
+            invalid_cred = LibGit2.UserPasswordCredentials(invalid_username, invalid_password)
 
             function gen_ex(; cached_cred=nothing)
                 quote
@@ -2076,7 +2073,7 @@ mktempdir() do dir
             expect_ssh_ex = quote
                 include($LIBGIT2_HELPER_PATH)
                 valid_cred = LibGit2.UserPasswordCredentials("foo", "bar")
-                payload = CredentialPayload(valid_cred)
+                payload = CredentialPayload(valid_cred, allow_ssh_agent=false)
                 credential_loop(valid_cred, "ssh://github.com/repo", Nullable(""),
                     Cuint(LibGit2.Consts.CREDTYPE_SSH_KEY), payload)
             end
@@ -2089,7 +2086,7 @@ mktempdir() do dir
             expect_https_ex = quote
                 include($LIBGIT2_HELPER_PATH)
                 valid_cred = LibGit2.SSHCredentials("foo", "", "", "")
-                payload = CredentialPayload(valid_cred)
+                payload = CredentialPayload(valid_cred, allow_ssh_agent=false)
                 credential_loop(valid_cred, "https://github.com/repo", Nullable(""),
                     Cuint(LibGit2.Consts.CREDTYPE_USERPASS_PLAINTEXT), payload)
             end
@@ -2109,7 +2106,7 @@ mktempdir() do dir
             ex = quote
                 include($LIBGIT2_HELPER_PATH)
                 valid_cred = LibGit2.UserPasswordCredentials("foo", "bar")
-                payload = CredentialPayload(valid_cred)
+                payload = CredentialPayload(valid_cred, allow_ssh_agent=false)
                 credential_loop(valid_cred, "foo://github.com/repo", Nullable(""),
                     $allowed_types, payload)
             end